adding premium plan and vnpay payment

+++ Backend/koi_care_system/src/main/java/com/swpproject/koi_care_system/dto/PaymentDto.java
@@ -1,5 +1,6 @@
 package com.swpproject.koi_care_system.dto;
 
+import com.swpproject.koi_care_system.enums.OrderStatus;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -7,7 +8,7 @@ import java.time.LocalDateTime;
 @Data
 public class PaymentDto {
     private LocalDateTime createDate;
-    private Double amount;
+    private Long amount;
     private String status;
     private String invoiceCode;
     private String transactionCode;
